Swift add tests for manifest
