+++ .gitignore
@@ -17,3 +17,4 @@ build_rs_cov.profraw
 cobertura.xml
 jna-*.jar
 .swiftpm/*
+default.profraw
