+++ Package.swift
@@ -57,7 +57,7 @@ let package = Package(
 				.target(name: "Sargon"),
 				.product(name: "CustomDump", package: "swift-custom-dump"),
 			],
-			path: "apple/Tests/SargonTests"
+			path: "apple/Tests"
 		),
 	]
 )
