+++ apple/Tests/Utils/EntityProtocolTests.swift
